% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{approx.own}
\alias{approx.own}
\title{Interpolation}
\usage{
approx.own(x, y, xout, MARGIN = NULL)
}
\arguments{
\item{x}{numeric vectors giving the coordinates of the points to be
    interpolated.  Alternatively a single plotting structure can be
    specified: see \code{\link{xy.coords}}.}

\item{y}{numeric vectors giving the coordinates of the points to be
    interpolated.  Alternatively a single plotting structure can be
    specified: see \code{\link{xy.coords}}.}

\item{xout}{an optional set of numeric values specifying where
    interpolation is to take place.}

\item{MARGIN}{If y has dimensions, i.e. is a matrix, over which of them should the function be applied? See also \code{\link[base]{apply}}.}
}
\description{
Approximates also then, when xout is below/above the minimum/maximum value given in x.
}
\author{
Daniel Hoop
}
\keyword{internal}
