% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{calcMean}
\alias{calcMean}
\title{Calculate means for different index levels}
\usage{
calcMean(
  data,
  cols = NULL,
  weights = NULL,
  indices,
  indexOrder,
  contIndex = NULL,
  nObsInContIndexClass = 3,
  maxContIndexClasses = 10,
  na.rm = TRUE,
  countNonNa = FALSE,
  truncProbs = NULL,
  type = c("mean", "median")
)
}
\arguments{
\item{data}{The data.frame containing the columns to be calculated.}

\item{cols}{The columns in \code{data} for which the means should be caluclated. Character vector. If empty, then the means for all columns in \code{data} will be calculated.}

\item{weights}{The weights for each observation in \code{data}. Character vector. If empty, then unweighted means will be calculated.}

\item{indices}{The indices for which the means should be calculated. A data.frame (or list). See the function tapply as a rough reference.}

\item{indexOrder}{The index combinations for which the average values should be calculated. A list containing the combination of indices in each list place.}

\item{contIndex}{A continous index, e.g. the size of a farm. Numeric vector.}

\item{nObsInContIndexClass}{The desired number of observations in each class that will be build for the \code{contIndex}.}

\item{maxContIndexClasses}{The maximum number of classes that will be built for \code{contIndex}.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the average values are getting calculated.}

\item{countNonNa}{A logical value indicating whether only the number of observations in a strata should be counted (\code{countNonNa = FALSE}) or,}

\item{truncProbs}{The probability of quantiles within which the truncated mean should be calculated. Numeric vector of length two. Both numbers must lie between 0 and 1 (inclusive).}

\item{type}{Either \code{"mean"} or \code{"median"} indicating if the (weighted) means or the (weighted) medians should be calculated.
in addition, for each column in each strata the number of non NA values should be counted (\code{countNonNa = TRUE}), which is computationally more expensive.}
}
\value{
A function to call in order to get the means for indices. See function \code{\link{getMean}} for more information.
}
\description{
Calculates the average values of each index level, the combination of which are specified.
}
\examples{
data <- data.frame(
  a=11:41,
  b=61:91)
indices <- data.frame(
  ind1=c( 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3,  4),
  ind2=c( 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 8, 8, 8, 9, 7, 7, 7, 8, 8, 7, 7, 7, 7, 8, 8, 8, 8, 9, 9, 9, 10),
  all =c( 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,  1))
indexOrder <- list(
  c('ind1', 'ind2'),
  'ind1',
  'all')
getMean <- calcMean(data = data, indices = indices, indexOrder = indexOrder)
print(getMean(raw = TRUE)$mean)
print(getMean(raw = TRUE)$n)
}
\author{
Daniel Hoop
}
