% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{charColsToNum}
\alias{charColsToNum}
\title{Change character columns to numeric/integer, if they can be coerced as such.}
\usage{
charColsToNum(x, stringsAsFactors = FALSE)
}
\arguments{
\item{x}{The data.frame/matrix to be converted.}

\item{stringsAsFactors}{Logical value indicating if character vectors should be converted to factors. Default is \code{FALSE}.}
}
\value{
A data.frame with same dimensions as \code{x}. Colnames will be equal but rownames might have changed.
}
\description{
Change character columns to numeric/integer, if they can be coerced as such.
}
\author{
Daniel Hoop
}
\keyword{internal}
