% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{checkExprValidity}
\alias{checkExprValidity}
\title{Checks expressions for errors}
\usage{
checkExprValidity(
  expr,
  dataFill = NULL,
  assignOp = c("<-", "="),
  assignmentForbidden = NULL,
  outputFile = NULL
)
}
\arguments{
\item{expr}{The expression to check for errors.}

\item{dataFill}{Optional data.frame in which the expression should be evaluated later on. If not NULL, then the function will check if variables that are being assigned in the expression do not already exist in dataFill. If they already exist, an error will be shown.}

\item{assignOp}{The assignment operator which assigns the values of the right hand side (RHS) to the left hand side (LHS). A character vector of any length, e.g. c("<-", "=")}

\item{assignmentForbidden}{Optional character vector that contains variables that should not be assigned.}

\item{outputFile}{Optional: Character or connection to a file. If circular references were found, all problematic lines will be written into that file.}
}
\value{
invisible(NULL) if everything is okay. Will cause an error if problems were detected in the expression.
}
\description{
Checks for circular references, erroneouss +/- calculations over several lines, and more (see arguments).
}
\examples{
checkExprValidity(quote({a <- 1})) # Nothing happens. Everything is ok.
checkExprValidity(quote({
  a <- 1
  + 2
}))
# An error will be shown because a new line must not start with a '+' or '-' sign.
# For more examples, see also the function 'findCircularReferences'.
}
\seealso{
\code{\link{findCircularReferences}}
}
\author{
Daniel Hoop
}
