% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{combineQuotes}
\alias{combineQuotes}
\title{Combine quoted expressions}
\usage{
combineQuotes(...)
}
\arguments{
\item{...}{Either all quoted expressions that should be combined or a list containing all of them.}
}
\value{
A quoted expressions that contains all expressions given in \code{...}.
}
\description{
Combine quoted expressions
}
\examples{
# All given examples yield the same result.
combineQuotes(
  quote({a <- 1})
  quote({b <- 2})
  quote({c <- 3}))
combineQuotes(
  list(quote({a <- 1}),
       quote({b <- 2}),
       quote({c <- 3})))
combineQuotes(
  quote({a <- 1}),
  c("b <- 2"
    quote({c <- 3})))
}
\author{
Daniel Hoop
}
