% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{createLookupWithMeanMethod}
\alias{createLookupWithMeanMethod}
\title{Create the lookup function}
\usage{
createLookupWithMeanMethod(
  ref,
  weights = NULL,
  type = c("mean", "median"),
  ...
)
}
\arguments{
\item{ref}{The reference data set from which the average values should be derived (data.frame).}

\item{weights}{The weights for the observations given in the reference data set (numeric). If \code{is.null(weights)}, then all observations are weighted equally.}

\item{...}{Not used but needed such that the same parameters can be passed into different functions such as \code{\link{createLookupWithMeanMethod}} and \code{\link{distToWeight}}.}
}
\value{
The lookup function which will have the following arguments:\cr
\code{na.rm} Logical value indicating if na.rm should be applied when calculating the mean value of the reference group.\cr
\code{zeroToNaBeforeCalc} Logical value indicating if zero values should be turned to NA before calculating the mean of the reference group.\cr
\code{zeroToNaBeforeCalc} Logical value indicating if NA values should be turned to 0, when the result is returned.\cr
\code{getRefForDebug} Logical value indicating the whole reference dataset should be returned instead of a calculated average values.
This is useful if you are browsing the calculation expression and want to check how the reference dataset looks for a specific observation.
}
\description{
Creates the lookup function to be used in the argument \code{expr} of the function \code{\link{imputeDomainDriven}}.
}
\author{
Daniel Hoop
}
