% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{.getIdFromInd}
\alias{.getIdFromInd}
\title{A function to convert ind to ids. Helper used in `getNearestObs`.}
\usage{
.getIdFromInd(ind, id)
}
\description{
A function to convert ind to ids. Helper used in `getNearestObs`.
}
\examples{
ind <- matrix(sample(1:10, 20, replace = TRUE), ncol = 2)
id <- letters[1:10]
.getIdFromInd(ind, id)
}
\author{
Daniel Hoop
}
\keyword{internal}
