% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{.splitExceptInBracket}
\alias{.splitExceptInBracket}
\title{Splits an expression where operators occur, but not inside brackets.}
\usage{
.splitExceptInBracket(
  txt,
  operators = NULL,
  notAroundOperator = NULL,
  ifIsAnOperator = TRUE,
  splitIfElseOnly = FALSE
)
}
\description{
Splits an expression where operators occur, but not inside brackets.
}
\examples{
.splitExceptInBracket("(a+b+c)+1++1+")
[1] "(a+b+c)" "1"       ""        "1"
.splitExceptInBracket("a+b", splitIfElseOnly=TRUE)
[1] "a+b"
.splitExceptInBracket("1 <- if (a == b) c else d", splitIfElseOnly=TRUE)
[1] "1 <- if (a == b)" "c"                "d"
.splitExceptInBracket(c("a = b+b", "a <-b+b"), operators = c("=", "<-"))
[[1]]
[1] "a"   "b+b"
[[2]]
[1] "a"   "b+b"
.splitExceptInBracket("a <= b", operators = c("<-", "="), notAroundOperator = c("<", ">", "!", "="))
}
\author{
Daniel Hoop
}
\keyword{internal}
