% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{getMean}
\alias{getMean}
\title{Imputute by means of strata averages}
\usage{
getMean(
  index,
  cols = NULL,
  nMin,
  count = c("all", "nonNa"),
  forceValue = FALSE,
  raw = FALSE
)
}
\arguments{
\item{index}{The index for each observation of which the mean value should be looked up.}

\item{cols}{The colums of which the average values should be looked up.}

\item{nMin}{The minimum number of observations that must be 'behind' a mean value in \code{means}, in order for it to be valuable.\cr
The indices will be 'traversed' in the indexOrder that was specified to create \code{means} by using the function \code{\link{calcMean}}.\cr
If in the first index combination specified in 'indexOrder' not enough observations are found, then the next higher index combination will be chosen, etc.\cr}

\item{count}{Either \code{"all"} or \code{"nonNa"}, indicating if, for the minimum number of observations given in \code{nMin}, all observations should be counted, or only those with non NA values.}

\item{forceValue}{Logical value indicating if the value of the last index according to the \code{indexOrder} given in \code{\link{calcMean}} should be taken, even if it does not provide the minimum required observations according to \code{nMin}.}

\item{raw}{Logical value indicating if the raw data should be returned instead of matching means to according index values.}

\item{contIndex}{The continuous index for each observation of which the mean value should be looked up.}
}
\value{
A list containing two list places 'mean' and 'index'.\cr
The list place 'mean' contains the means that were looked up.\cr
The list place 'index' shows which index combination was chosen to look up the mean.
}
\description{
Draws from pre-calculated average values but only if the number of observations in the strata is large enough.
}
\details{
If no index combination was found that offers enough observations, and \code{forceValue = FALSE} (default), then NA values will be returned for the respective rows in \code{index}.
}
\examples{
# First, pre-calculate the means
data <- data.frame(
  a=11:41,
  b=61:91)
indices <- data.frame(
  ind1=c( 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3,  4),
  ind2=c( 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 8, 8, 8, 9, 7, 7, 7, 8, 8, 7, 7, 7, 7, 8, 8, 8, 8, 9, 9, 9, 10),
  all =c( 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,  1))
indexOrder <- list(
  c('ind1', 'ind2'),
  'ind1',
  'all')
contIndex <-
       c( 1, 2, 3, 4, 5, 6, 7, 8, 9,10, 1, 2, 3, 1,11,12,13, 1, 1, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3,  1)
getMean <- calcMean(data = data, cols=c("a","b","I(a/b)"), indices = indices, indexOrder = indexOrder, contIndex = contIndex)
print(getMean(raw = TRUE)$mean)
print(getMean(raw = TRUE)$n)

# Define the index of some observations that should be looked up.
index <- data.frame(
  ind1=c(1, 2, 3, 4),
  ind2=c(7, 8, 9, 10),
  all=c(0, 0, 0, 1))

# Now look up the means for some observations.
# Notice that the fourth observation did not get a value because it does not satisfy nMin.
getMean(index, nMin=10)
# The following line will get an imputed value for all observations because forceValue=TRUE.
getMean(index, nMin=10, forceValue=TRUE)
# If you give a continuous index, the results will be caluculated along that continuous index.
getMean(index, contIndex=c(1,2,NA,9), nMin=0)
}
\author{
Daniel Hoop
}
