% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{getMeanCalib}
\alias{getMeanCalib}
\title{Imputation by means of calibrated averages.}
\usage{
getMeanCalib(
  dataFill,
  dataGet,
  colsCalib,
  colsCalc,
  weightsFill = NULL,
  initialWeightsGet = NULL,
  na.rm = TRUE
)
}
\description{
Calibrates certain columns of the reference dataset \code{dataGet} to fit to the sum in \code{dataFill}. Yeat no detailed description available.
}
\author{
Daniel Hoop
}
\keyword{internal}
