% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{installFromCRAN}
\alias{installFromCRAN}
\title{Install a package from CRAN, but only if it is not yet installed}
\usage{
installFromCRAN(pkgs, ...)
}
\arguments{
\item{...}{Further arguments passed into \code{\link[utils:install.packages]{utils::install.packages}}.}

\item{x}{The name(s) of the package(s).}
}
\value{
Invisible \code{NULL}
}
\description{
Install a package from CRAN, but only if it is not yet installed
}
\author{
Daniel Hoop
}
\keyword{internal}
