% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{multiCorePost}
\alias{multiCorePost}
\title{Multicore post operations}
\usage{
multiCorePost(parClust = NULL)
}
\description{
Stops the parallel cluster given in optional argument \code{parClust} (if on Windows operating systems).
}
\author{
Daniel Hoop
}
\keyword{internal}
