% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{multiCorePrep}
\alias{multiCorePrep}
\title{Multicore preparations}
\usage{
multiCorePrep(nCores)
}
\arguments{
\item{nCores}{The number of cores to register.}
}
\description{
Installs required packages, registers a parallel cluster and returns it as value, if on Windows operating systems.
}
\author{
Daniel Hoop
}
\keyword{internal}
