% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgImpute.R
\name{quoteToString}
\alias{quoteToString}
\title{Converts a quoted expression to a string.}
\usage{
quoteToString(x, collapseTxt = TRUE)
}
\arguments{
\item{x}{The quoted expression to convert into a string.}

\item{collapseTxt}{Logical value indicating if the resulting character vector should be collapsed with newline characters \code{"\n"}.}
}
\description{
Converts a quoted expression to a string.
}
\author{
Daniel Hoop
}
\keyword{internal}
